fix_4016 Move input data thread to the BasicPerformance abstract class
